+++ kernel/include/kernel/pmm.h
@@ -4,10 +4,10 @@
 #include <stdint.h>
 #include <stddef.h>
 
-void pmm_initialize(uint32_t mem_upper_limit, uint32_t kernel_end);
-uint32_t pmm_alloc_block();
-void pmm_free_block(uint32_t block_addr);
-void pmm_deinit_region(uint32_t base, uint32_t size);
-void pmm_init_region(uint32_t base, uint32_t size);
+void pmm_initialize(uint64_t mem_upper_limit, uint64_t kernel_end);
+uint64_t pmm_alloc_block();
+void pmm_free_block(uint64_t block_addr);
+void pmm_deinit_region(uint64_t base, uint64_t size);
+void pmm_init_region(uint64_t base, uint64_t size);
 
 #endif
