+++ kernel/kernel/pmm.c
@@ -6,9 +6,9 @@
 #define BLOCK_SIZE 4096
 #define BLOCKS_PER_BYTE 0
 
-static uint32_t* pmm_bitmap = 0;
-static uint32_t used_blocks = 0;
-static uint32_t max_blocks = 0;
+static uint64_t* pmm_bitmap = 0;
+static uint64_t used_blocks = 0;
+static uint64_t max_blocks = 0;
 
 void mmap_set(int bit) {
 	pmm_bitmap[bit / 32] |= (1 << (bit % 32));
@@ -22,18 +22,18 @@ int mmap_test(int bit) {
 	return pmm_bitmap[bit / 32] & (1 << (bit % 32));
 }
 
-void pmm_initialize(uint32_t mem_size, uint32_t bitmap_start) {
+void pmm_initialize(uint64_t mem_size, uint64_t bitmap_start) {
 	max_blocks = mem_size / BLOCK_SIZE;
 	used_blocks = max_blocks;
 
 	// put the bitmap right where the kernel ends
-	pmm_bitmap = (uint32_t*)bitmap_start;
+	pmm_bitmap = (uint64_t*)bitmap_start;
 
 	memset(pmm_bitmap, 0xFF, max_blocks / 8);
 	printf("PMM: Bitmap initialized at 0x%x, tracking %d blocks.\n", bitmap_start, max_blocks);
 }
 
-void pmm_init_region(uint32_t base, uint32_t size) {
+void pmm_init_region(uint64_t base, uint64_t size) {
 	int align = base / BLOCK_SIZE;
 	int blocks = size / BLOCK_SIZE;
 
@@ -45,7 +45,7 @@ void pmm_init_region(uint32_t base, uint32_t size) {
 	mmap_set(0);
 }
 
-void pmm_deinit_region(uint32_t base, uint32_t size) {
+void pmm_deinit_region(uint64_t base, uint64_t size) {
 	int align = base / BLOCK_SIZE;
 	int blocks = size / BLOCK_SIZE;
 
@@ -55,8 +55,8 @@ void pmm_deinit_region(uint32_t base, uint32_t size) {
 	}
 }
 
-uint32_t pmm_alloc_block() {
-	for (uint32_t i = 0; i < max_blocks; i++) {
+uint64_t pmm_alloc_block() {
+	for (uint64_t i = 0; i < max_blocks; i++) {
 		if (!mmap_test(i)) {
 			mmap_set(i); // mark as read
 			used_blocks++;
@@ -67,8 +67,8 @@ uint32_t pmm_alloc_block() {
 	return 0;
 }
 
-void pmm_free_block(uint32_t addr) {
-	uint32_t block = addr / BLOCK_SIZE;
+void pmm_free_block(uint64_t addr) {
+	uint64_t block = addr / BLOCK_SIZE;
 	mmap_unset(block);
 	used_blocks--;
 }
